Hide non-configured matching criteria

+++ apps/modernization-ui/src/apps/deduplication/api/useMatchConfiguration.ts
@@ -4,8 +4,30 @@ import { Config } from 'config';
 
 export const useMatchConfiguration = () => {
     const [passes, setPasses] = useState<Pass[]>([]);
+    const [selectedPass, setSelectedPass] = useState<Pass | undefined>();
     const [error, setError] = useState<string | undefined>();
     const [loading, setLoading] = useState<boolean>(false);
+
+    const selectPass = (pass?: Pass) => {
+        // a new, unsaved pass was previously selected. Remove it from list
+        if (selectedPass !== undefined && selectedPass.id === undefined) {
+            setPasses(passes.filter((p) => p.id !== undefined));
+        }
+
+        setSelectedPass(pass);
+    };
+
+    const addPass = () => {
+        // adds a new pass to the beginning of the pass list
+        const newPass: Pass = {
+            name: 'New pass configuration',
+            blockingCriteria: [],
+            matchingCriteria: [],
+            active: false
+        };
+        setPasses([newPass, ...passes.filter((p) => p.id !== undefined)]);
+        setSelectedPass(newPass);
+    };
 
     const fetchConfiguration = async () => {
         setLoading(true);
@@ -19,7 +41,10 @@ export const useMatchConfiguration = () => {
             .then((response) => {
                 response
                     .json()
-                    .then((algorithm) => setPasses(algorithm.passes))
+                    .then((algorithm) => {
+                        setSelectedPass(undefined);
+                        setPasses(algorithm.passes);
+                    })
                     .catch(() => {
                         console.error('Failed to extract json for pass configuration.');
                     });
@@ -32,38 +57,70 @@ export const useMatchConfiguration = () => {
         setLoading(false);
     };
 
-    const deletePass = (id: number, successCallback?: () => void) => {
-        fetch(`${Config.deduplicationUrl}/api/configuration/pass/${id}`, {
-            method: 'DELETE',
-            headers: {
-                Accept: 'application/json'
-            }
-        })
-            .then((response) => {
-                response
-                    .json()
-                    .then((algorithm) => setPasses(algorithm.passes))
-                    .catch(() => {
-                        console.error('Failed to extract json for pass configuration.');
-                    });
+    const deletePass = (id?: number, successCallback?: () => void) => {
+        setLoading(true);
+        // are we deleting the unsaved new pass
+        if (id === undefined) {
+            setPasses(passes.filter((p) => p.id !== undefined));
+            successCallback?.();
+        } else {
+            fetch(`${Config.deduplicationUrl}/api/configuration/pass/${id}`, {
+                method: 'DELETE',
+                headers: {
+                    Accept: 'application/json'
+                }
             })
-            .catch((error) => {
-                console.error(error);
-                setError('Failed to delete pass');
-            });
+                .then((response) => {
+                    response
+                        .json()
+                        .then((algorithm) => setPasses(algorithm.passes))
+                        .catch(() => {
+                            console.error('Failed to extract json for pass configuration.');
+                        });
+                })
+                .catch((error) => {
+                    console.error(error);
+                    setError('Failed to delete pass');
+                });
 
-        successCallback?.();
+            successCallback?.();
+        }
+        setSelectedPass(undefined);
+        setLoading(false);
     };
 
-    const savePass = (pass: Pass, successCallback?: () => void) => {
+    const savePass = (pass: Pass, successCallback?: (passName: string) => void) => {
+        setLoading(true);
         if (pass.id) {
             updatePass(pass, successCallback);
         } else {
             createPass(pass, successCallback);
         }
+        setLoading(false);
     };
 
-    const updatePass = (pass: Pass, successCallback?: () => void) => {
+    const updatePass = (pass: Pass, successCallback?: (passName: string) => void) => {
+        makeUpdateRequest(pass, (passes) => {
+            setPasses(passes);
+            const updatedPass = passes.find((p) => p.id === pass.id);
+            setSelectedPass(updatedPass);
+            successCallback?.(updatedPass?.name ?? '');
+        });
+    };
+
+    const updatePassName = (pass: Pass, successCallback?: (passName: string) => void) => {
+        makeUpdateRequest(pass, (passes) => {
+            if (selectedPass !== undefined && selectedPass.id === undefined) {
+                setPasses([selectedPass, ...passes]);
+            } else {
+                setPasses(passes);
+            }
+            const updatedPass = passes.find((p) => p.id === pass.id);
+            successCallback?.(updatedPass?.name ?? '');
+        });
+    };
+
+    const makeUpdateRequest = (pass: Pass, onResponse: (passes: Pass[]) => void) => {
         fetch(`${Config.deduplicationUrl}/api/configuration/pass/${pass.id}`, {
             method: 'PUT',
             headers: {
@@ -75,7 +132,7 @@ export const useMatchConfiguration = () => {
             .then((response) => {
                 response
                     .json()
-                    .then((algorithm) => setPasses(algorithm.passes))
+                    .then((algorithm: { passes: Pass[] }) => onResponse(algorithm.passes))
                     .catch(() => {
                         console.error('Failed to extract json for pass configuration.');
                     });
@@ -84,11 +141,18 @@ export const useMatchConfiguration = () => {
                 console.error(error);
                 setError('Failed to save pass');
             });
-
-        successCallback?.();
     };
 
-    const createPass = (pass: Pass, successCallback?: () => void) => {
+    const createPass = (pass: Pass, successCallback?: (passName: string) => void) => {
+        makeCreateRequest(pass, (passes) => {
+            setPasses(passes);
+            const newPass = passes[passes.length - 1];
+            setSelectedPass(newPass);
+            successCallback?.(newPass.name);
+        });
+    };
+
+    const makeCreateRequest = (pass: Pass, onResponse: (passes: Pass[]) => void) => {
         fetch(`${Config.deduplicationUrl}/api/configuration/pass`, {
             method: 'POST',
             headers: {
@@ -100,7 +164,7 @@ export const useMatchConfiguration = () => {
             .then((response) => {
                 response
                     .json()
-                    .then((algorithm) => setPasses(algorithm.passes))
+                    .then((algorithm: { passes: Pass[] }) => onResponse(algorithm.passes))
                     .catch(() => {
                         console.error('Failed to extract json for pass configuration.');
                     });
@@ -109,13 +173,22 @@ export const useMatchConfiguration = () => {
                 console.error(error);
                 setError('Failed to save pass');
             });
-
-        successCallback?.();
     };
 
     useEffect(() => {
         fetchConfiguration();
     }, []);
 
-    return { fetchConfiguration, deletePass, savePass, loading, error, passes };
+    return {
+        loading,
+        error,
+        passes,
+        selectedPass,
+        fetchConfiguration,
+        deletePass,
+        savePass,
+        updatePassName,
+        selectPass,
+        addPass
+    };
 };
